Add GUI
Removed the old logic of 5 rounds and reworked the battle function.
I decided to make the whole GUI before committing, since it was
something to be done in one sitting. There is a button for each player
selection, and making the selection triggers the round. The buttons are
disabled, the text of each selection on the upper div is changed based
on the selection and a little animation to make the words clash plays.
After that, the result is displayed, the scores updated and a button to
start a new round appears. This button re-enables the selection buttons
and remove the animations and other added styles.